NEW component folder added for components to sit in || files paths updated across app for new locations
