+++ app/components/cards/cards.component.ts
@@ -1,9 +1,9 @@
 import {Component} from '@angular/core';
-import {testDirective} from '../directives/form.directive';
+// import {testDirective} from '../../directives/form.directive';
 
 @Component({
     selector: 'app-cards',
-    templateUrl: './app/cards/cards.template.html',
+    templateUrl: './app/components/cards/cards.template.html',
     styleUrls: ['../css/cards/cards.css']
 })
 export class CardsComponent {
