+++ app/components/event/event-template.html
