+++ app/components/event/event.scss
