+++ app/components/home/home.component.ts
@@ -2,7 +2,7 @@ import {Component} from '@angular/core';
 
 @Component({
     selector: 'app-home',
-    templateUrl: './app/home/home-template.html'
+    templateUrl: './app/components/home/home-template.html'
 })
 export class HomeComponent {
     welcome : string;
